svg/painting/parsing/fill-opacity-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: getComputedStyle().fillOpacity</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#FillOpacity" rel="help"/><h:meta content="fill-opacity computed value is clamped to the range [0,1]." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("fill-opacity", "-1", "0");
test_computed_value("fill-opacity", "0.5");
test_computed_value("fill-opacity", "3", "1");
test_computed_value("fill-opacity", "-100%", "0");
test_computed_value("fill-opacity", "50%", "0.5");
test_computed_value("fill-opacity", "300%", "1");

  ]]></script></svg>
